requirements.txt
# Core Browser Automation (2026)
playwright==1.48.0
playwright-stealth==1.0.6

# Configuration & Data
pyyaml==6.0.1
pydantic==2.9.2
python-dotenv==1.0.1

# Async HTTP & Networking
httpx==0.27.2
aiofiles==24.1.0

# Email Services (ALL FREE)
pymailtm==1.1.1

# Utilities
pyperclip==1.9.0
tqdm==4.67.0
rich==13.9.2
tenacity==9.0.0
fake-useragent==1.5.1
loguru==0.7.2

# Note for Termux Users:
# Some dependencies may require compilation. Install these first:
# pkg install python nodejs rust binutils clang
# pip install --upgrade pip setuptools wheel
